Give banners their own viewmodels (#1768)
