+++ tools/Dashboard/DevHome.Dashboard/Views/DashboardView.xaml
@@ -60,10 +60,10 @@
                 <!-- Top Banner - Default/First run experience - shown until user dismissed -->
                 <commonviews:Banner x:Name="DashboardBanner"
                     TextWidth="450"
-                    Visibility="{x:Bind ViewModel.ShowDashboardBanner,Mode=OneWay}"
+                    Visibility="{x:Bind BannerViewModel.ShowDashboardBanner,Mode=OneWay}"
                     HideButtonVisibility="true"
-                    HideButtonCommand="{x:Bind ViewModel.HideDashboardBannerButtonCommand,Mode=OneWay}"
-                    ButtonCommand="{x:Bind ViewModel.DashboardBannerButtonCommand,Mode=OneWay}"
+                    HideButtonCommand="{x:Bind BannerViewModel.HideDashboardBannerButtonCommand,Mode=OneWay}"
+                    ButtonCommand="{x:Bind BannerViewModel.DashboardBannerButtonCommand,Mode=OneWay}"
                     x:Uid="ms-resource:///DevHome.Dashboard/Resources/DashboardBanner"
                     BackgroundSource="{ThemeResource DashboardBannerBack}"
                     OverlaySource="{ThemeResource DashboardBannerFront}"
