+++ tools/ExtensionLibrary/DevHome.ExtensionLibrary/Extensions/ServiceExtensions.cs
@@ -6,12 +6,14 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace DevHome.ExtensionLibrary.Extensions;
+
 public static class ServiceExtensions
 {
     public static IServiceCollection AddExtensionLibrary(this IServiceCollection services, HostBuilderContext context)
     {
         // View-models
         services.AddTransient<ExtensionLibraryViewModel>();
+        services.AddTransient<ExtensionLibraryBannerViewModel>();
 
         return services;
     }
